Application: added option to kill application with a custom "kill_cmd"

+++ infrastructure/platform/linux/Applications/Application.py
@@ -1,23 +1,26 @@
 import time
 
 class _Application(object):
-    def __init__(self, cmd, ui, app_cmd, app_title):
+    def __init__(self, cmd, ui, start_cmd, stop_cmd = None):
         self._cmd = cmd
         self._ldtp = ui.ldtp
         self._dogtail = ui.dogtail
-        self._app_cmd = app_cmd
-        self._app_title = app_title
+        self._start_cmd = start_cmd
+        self._stop_cmd = stop_cmd
         self._process = None
 
     def start(self):
-        self._process = self._cmd(self._app_cmd)
+        self._process = self._cmd(self._start_cmd)
 
     def stop(self):
         if self._process:
-            self._process.terminate()
-            time.sleep(2)
-            if (0 != self._cmd('ls /proc/' + str(self._process.pid)).wait()): # if not killed
-                self._process.kill()
+            if self._stop_cmd:
+                self._cmd(self._stop_cmd)
+            else:
+                self._process.terminate()
+                time.sleep(2)
+                if (0 != self._cmd('ls /proc/' + str(self._process.pid)).wait()): # if not killed
+                    self._process.kill()
 
     def find_children(self, app, name = None, roleName = None):
         children = app.findChildren(self._dogtail.predicate.GenericPredicate(name, roleName))
@@ -25,5 +28,5 @@ class _Application(object):
 
 class _Editor(_Application):
     def start(self, doc = ''):
-        self._app_cmd = self._app_cmd + ' ' + doc
+        self._start_cmd = self._start_cmd + ' ' + doc
         super(_Editor, self).start()
